вынес логику отображения точки маршрута в отдельный презентер
