+++ src/view/point-view.js
@@ -64,21 +64,21 @@ function createPointTemplate(point, destination, offers) {
 export default class PointView extends AbstractView {
   #point = null;
   #destination = null;
-  #offers = null;
+  #checkedOffers = null;
   #handleEditClick = null;
 
-  constructor({point, destination, offers, onEditClick}) {
+  constructor({point, destination, checkedOffers, onEditClick}) {
     super();
     this.#point = point;
     this.#destination = destination;
-    this.#offers = offers;
+    this.#checkedOffers = checkedOffers;
     this.#handleEditClick = onEditClick;
 
     this.element.querySelector('.event__rollup-btn').addEventListener('click', this.#editClickHandler);
   }
 
   get template() {
-    return createPointTemplate(this.#point, this.#destination, this.#offers);
+    return createPointTemplate(this.#point, this.#destination, this.#checkedOffers);
   }
 
   #editClickHandler = (evt) => {
